feat(web): extract MCP config validator lib and expand redaction tests
Move secret redaction, transport classification, and server validation
into a pure library module with focused unit tests so the validator
wrapper stays thin and patch coverage stays high.
