+++ apps/web/src/lib/mcp-config-validator.ts
@@ -1,473 +1,13 @@
-import { isPinnedPackageSpec, parsePackageSpec } from "@heyclaude/registry/package-spec";
+import {
+  buildReportText,
+  extractServers,
+  isRecord,
+  sanitizeConfigValue,
+  validateServer,
+  type McpConfigValidation,
+} from "@/lib/mcp-config-validator-lib";
 
-export type McpConfigServerReport = {
-  name: string;
-  transport: "stdio" | "remote" | "unknown";
-  command?: string;
-  url?: string;
-  packageName?: string;
-  envKeys: string[];
-  errors: string[];
-  warnings: string[];
-};
-
-export type McpConfigValidation = {
-  ok: boolean;
-  errors: string[];
-  warnings: string[];
-  servers: McpConfigServerReport[];
-  fixedConfigText: string;
-  reportText: string;
-  redactedSecretCount: number;
-};
-
-const SENSITIVE_ENV_PATTERN =
-  /(api[_-]?key|auth|authorization|bearer|client[_-]?secret|credential|env|password|private[_-]?key|secret|token|x-api-key)/i;
-const PLACEHOLDER_PATTERN = /(\$\{[A-Z0-9_]+\}|YOUR_|REPLACE_|INSERT_|<[^>]+>|\bxxx+\b|\bTODO\b)/i;
-const SECRET_VALUE_PATTERN =
-  /\b(gh[pousr]_[A-Za-z0-9_]{20,}|github_pat_[A-Za-z0-9_]{40,}|glpat-[A-Za-z0-9_-]{20,}|sk-(?:proj-)?[A-Za-z0-9_-]{20,}|xox[baprs]-[A-Za-z0-9-]{20,}|AKIA[0-9A-Z]{16}|AIza[0-9A-Za-z_-]{20,}|Bearer\s+[A-Za-z0-9._~+/=-]{16,})\b/;
-const SHELL_OPERATOR_PATTERN = /(?:&&|\|\||[;|`<>]|\$\()/;
-const SENSITIVE_SPLIT_ARG_KEYS = new Set([
-  "api_key",
-  "apikey",
-  "auth",
-  "authorization",
-  "bearer",
-  "client_secret",
-  "clientsecret",
-  "password",
-  "private_key",
-  "privatekey",
-  "secret",
-  "token",
-  "x_api_key",
-  "xapikey",
-]);
-
-type RedactedUrlPlaceholder =
-  | { kind: "username"; placeholder: string }
-  | { kind: "password"; placeholder: string }
-  | { kind: "query"; key: string; placeholder: string };
-
-function escapeRegExp(value: string) {
-  return value.replace(/[.*+?^${}()|[\]\\]/g, "\\$&");
-}
-
-function encodeSearchParamValue(value: string) {
-  return new URLSearchParams([["", value]]).toString().slice(1);
-}
-
-function encodeSearchParamKey(value: string) {
-  return new URLSearchParams([[value, ""]]).toString().slice(0, -1);
-}
-
-function decodeRedactedPlaceholders(value: string, placeholders: RedactedUrlPlaceholder[]) {
-  return placeholders.reduce((decoded, redaction) => {
-    if (redaction.kind === "username") {
-      return decoded.replace("://$%7BURL_USERNAME%7D", `://${redaction.placeholder}`);
-    }
-    if (redaction.kind === "password") {
-      return decoded.replace(":$%7BURL_PASSWORD%7D@", `:${redaction.placeholder}@`);
-    }
-
-    const key = escapeRegExp(encodeSearchParamKey(redaction.key));
-    const placeholder = escapeRegExp(encodeSearchParamValue(redaction.placeholder));
-    return decoded.replace(
-      new RegExp(`([?&]${key}=)${placeholder}(?=([&#]|$))`),
-      (_match, prefix: string) => `${prefix}${redaction.placeholder}`,
-    );
-  }, value);
-}
-
-function isRecord(value: unknown): value is Record<string, unknown> {
-  return Boolean(value && typeof value === "object" && !Array.isArray(value));
-}
-
-function asStringArray(value: unknown) {
-  return Array.isArray(value)
-    ? value.filter((item): item is string => typeof item === "string")
-    : [];
-}
-
-function normalizeServerName(value: string) {
-  return value.trim();
-}
-
-function redactEnvValue(key: string, value: unknown) {
-  const normalized = String(value ?? "");
-  if (!SENSITIVE_ENV_PATTERN.test(key) && !SECRET_VALUE_PATTERN.test(normalized)) {
-    return normalized;
-  }
-  if (!normalized || PLACEHOLDER_PATTERN.test(normalized)) return normalized;
-  const placeholderKey = key || "SECRET";
-  return `\${${placeholderKey.toUpperCase().replace(/[^A-Z0-9_]/g, "_")}}`;
-}
-
-function redactUrlValue(value: string) {
-  try {
-    const parsed = new URL(value);
-    let redacted = false;
-    const redactedPlaceholders: RedactedUrlPlaceholder[] = [];
-    if (parsed.username) {
-      const placeholder = "${URL_USERNAME}";
-      parsed.username = placeholder;
-      redactedPlaceholders.push({ kind: "username", placeholder });
-      redacted = true;
-    }
-    if (parsed.password) {
-      const placeholder = "${URL_PASSWORD}";
-      parsed.password = placeholder;
-      redactedPlaceholders.push({ kind: "password", placeholder });
-      redacted = true;
-    }
-    for (const [key, queryValue] of parsed.searchParams.entries()) {
-      if (SENSITIVE_ENV_PATTERN.test(key) || SECRET_VALUE_PATTERN.test(queryValue)) {
-        const placeholder = `\${${key.toUpperCase().replace(/[^A-Z0-9_]/g, "_") || "SECRET"}}`;
-        parsed.searchParams.set(key, placeholder);
-        redactedPlaceholders.push({ kind: "query", key, placeholder });
-        redacted = true;
-      }
-    }
-    const serialized = decodeRedactedPlaceholders(parsed.toString(), redactedPlaceholders);
-    const fallback = redactEnvValue("url", serialized);
-    return {
-      value: fallback,
-      redactedCount: redacted || fallback !== serialized ? 1 : 0,
-    };
-  } catch {
-    const fallback = redactEnvValue("url", value);
-    return {
-      value: fallback,
-      redactedCount: fallback === value ? 0 : 1,
-    };
-  }
-}
-
-function redactArgValue(value: string) {
-  const normalized = value.trim();
-  if (!normalized) return { value, redactedCount: 0 };
-  if (/^https?:\/\//i.test(normalized)) return redactUrlValue(normalized);
-  const equalIndex = normalized.indexOf("=");
-  if (equalIndex > 0) {
-    const rawKey = normalized.slice(0, equalIndex);
-    const rawValue = normalized.slice(equalIndex + 1);
-    if (SENSITIVE_ENV_PATTERN.test(rawKey) || SECRET_VALUE_PATTERN.test(rawValue)) {
-      const placeholder = rawKey.toUpperCase().replace(/[^A-Z0-9_]/g, "_");
-      return {
-        value: `${rawKey}=\${${placeholder || "SECRET"}}`,
-        redactedCount: 1,
-      };
-    }
-  }
-  if (SECRET_VALUE_PATTERN.test(normalized)) {
-    return { value: "${SECRET}", redactedCount: 1 };
-  }
-  return { value, redactedCount: 0 };
-}
-
-function splitArgPlaceholder(value: string) {
-  const normalized = value.trim();
-  if (!normalized.startsWith("-") || normalized.includes("=")) return "";
-  const key = normalized
-    .replace(/^-+/, "")
-    .trim()
-    .toLowerCase()
-    .replace(/[^a-z0-9]+/g, "_")
-    .replace(/^_+|_+$/g, "");
-  return SENSITIVE_SPLIT_ARG_KEYS.has(key) ? key.toUpperCase().replace(/[^A-Z0-9_]/g, "_") : "";
-}
-
-function redactArgArray(values: unknown[]): SanitizedValue {
-  let redactedCount = 0;
-  let pendingPlaceholder = "";
-  const sanitizedItems = values.map((item) => {
-    if (typeof item !== "string") {
-      pendingPlaceholder = "";
-      const sanitized = sanitizeConfigValue("args", item);
-      redactedCount += sanitized.redactedCount;
-      return sanitized.value;
-    }
-
-    const normalized = item.trim();
-    if (
-      pendingPlaceholder &&
-      normalized &&
-      !normalized.startsWith("-") &&
-      !PLACEHOLDER_PATTERN.test(normalized)
-    ) {
-      const placeholder = pendingPlaceholder;
-      redactedCount += 1;
-      pendingPlaceholder = "";
-      return `\${${placeholder || "SECRET"}}`;
-    }
-
-    const sanitized = redactArgValue(item);
-    redactedCount += sanitized.redactedCount;
-    pendingPlaceholder = splitArgPlaceholder(item);
-    return sanitized.value;
-  });
-  return { value: sanitizedItems, redactedCount };
-}
-
-type SanitizedValue = {
-  value: unknown;
-  redactedCount: number;
-};
-
-function sanitizeConfigValue(key: string, value: unknown): SanitizedValue {
-  if (Array.isArray(value)) {
-    if (key.toLowerCase() === "args") return redactArgArray(value);
-    let redactedCount = 0;
-    const sanitizedItems = value.map((item) => {
-      const sanitized = sanitizeConfigValue(key, item);
-      redactedCount += sanitized.redactedCount;
-      return sanitized.value;
-    });
-    return { value: sanitizedItems, redactedCount };
-  }
-
-  if (isRecord(value)) {
-    let redactedCount = 0;
-    const entries = Object.entries(value).map(([entryKey, entryValue]) => {
-      const sanitized = sanitizeConfigValue(entryKey, entryValue);
-      redactedCount += sanitized.redactedCount;
-      return [entryKey, sanitized.value];
-    });
-    return { value: Object.fromEntries(entries), redactedCount };
-  }
-
-  const normalized = String(value ?? "");
-  if (typeof value === "string" && key.toLowerCase() === "url") {
-    return redactUrlValue(normalized);
-  }
-  if (typeof value === "string" && key.toLowerCase() === "args") {
-    return redactArgValue(normalized);
-  }
-  const redacted = redactEnvValue(key, normalized);
-  if (redacted !== normalized) {
-    return {
-      value: redacted,
-      redactedCount: 1,
-    };
-  }
-
-  return {
-    value,
-    redactedCount: 0,
-  };
-}
-
-// First package operand of a runner's args: the first token that is neither a
-// flag nor the `--` separator. The package is the first operand regardless of
-// where `--` sits, so this works for both `<runner> pkg -- args` (args follow
-// the package) and `<runner> -- pkg` (the separator precedes the package).
-function firstPackageOperand(args: string[]) {
-  return args.find((arg) => arg && arg !== "--" && !arg.startsWith("-")) || "";
-}
-
-function commandName(command: string) {
-  const base =
-    command
-      .trim()
-      .split(/[\\/]/)
-      .pop()
-      ?.replace(/^["']|["']$/g, "")
-      .toLowerCase() || "";
-  return base.replace(/\.(?:bat|cmd|com|exe|ps1)$/i, "");
-}
-
-function packageRunnerName(command: string, args: string[]) {
-  const lower = commandName(command);
-  if (["npx", "uvx", "bunx"].includes(lower)) return lower;
-  if ((lower === "pnpm" || lower === "yarn") && args.includes("dlx")) {
-    return `${lower} dlx`;
-  }
-  if (lower === "npm" && args.some((arg) => ["exec", "x"].includes(arg))) {
-    return "npm exec";
-  }
-  return "";
-}
-
-function packageFromRunner(command: string, args: string[]) {
-  const lower = commandName(command);
-  if (lower === "npx") return firstPackageOperand(args);
-  if (["uvx", "bunx"].includes(lower)) return firstPackageOperand(args);
-  if (lower === "pnpm" || lower === "yarn") {
-    const runnerIndex = args.findIndex((arg) => arg === "dlx");
-    return firstPackageOperand(runnerIndex >= 0 ? args.slice(runnerIndex + 1) : args);
-  }
-  if (lower === "npm") {
-    const runnerIndex = args.findIndex((arg) => ["exec", "x"].includes(arg));
-    return firstPackageOperand(runnerIndex >= 0 ? args.slice(runnerIndex + 1) : args);
-  }
-  if (lower === "docker") {
-    const runIndex = args.findIndex((arg) => arg === "run");
-    return runIndex >= 0 ? firstPackageOperand(args.slice(runIndex + 1)) : "";
-  }
-  return "";
-}
-
-function extractServers(payload: unknown) {
-  if (!isRecord(payload)) {
-    return { servers: {}, rootError: "Config must be a JSON object." };
-  }
-  if (isRecord(payload.mcpServers)) {
-    return { servers: payload.mcpServers, wrapped: false };
-  }
-  const directServerShape =
-    Object.keys(payload).length > 0 &&
-    Object.values(payload).every(
-      (value) => isRecord(value) && ("command" in value || "url" in value),
-    );
-  if (directServerShape) return { servers: payload, wrapped: true };
-  return {
-    servers: {},
-    rootError: "Config must include an mcpServers object.",
-  };
-}
-
-function validateServer(name: string, raw: unknown) {
-  const errors: string[] = [];
-  const warnings: string[] = [];
-  const normalizedName = normalizeServerName(name);
-
-  if (!/^[A-Za-z0-9][A-Za-z0-9._-]{0,63}$/.test(normalizedName)) {
-    errors.push(
-      "Server name must be 1-64 characters of letters, numbers, dot, underscore, or dash.",
-    );
-  }
-  if (!isRecord(raw)) {
-    return {
-      name: normalizedName || name,
-      transport: "unknown" as const,
-      envKeys: [],
-      errors: ["Server config must be an object."],
-      warnings,
-      sanitized: {},
-      redactedSecretCount: 0,
-    };
-  }
-
-  const hasCommand = "command" in raw;
-  const command = typeof raw.command === "string" ? raw.command.trim() : "";
-  const url = typeof raw.url === "string" ? raw.url.trim() : "";
-  const args = asStringArray(raw.args);
-  const env = isRecord(raw.env) ? raw.env : {};
-  const envKeys = Object.keys(env).sort();
-  const sanitizedRaw = sanitizeConfigValue("", raw);
-  const sanitizedRawValue = isRecord(sanitizedRaw.value) ? sanitizedRaw.value : {};
-  const sanitizedArgs = asStringArray(sanitizedRawValue.args);
-  const sanitized = {
-    ...sanitizedRawValue,
-    ...(command ? { command } : {}),
-  };
-  const redactedSecretCount = sanitizedRaw.redactedCount;
-
-  if (!command && !url) {
-    errors.push("Server must define command for stdio or url for remote transport.");
-  }
-  if (hasCommand && typeof raw.command !== "string") {
-    errors.push("command must be a string.");
-  }
-  if ("args" in raw && !Array.isArray(raw.args)) {
-    errors.push("args must be an array of strings.");
-  }
-  if ("args" in raw && Array.isArray(raw.args) && args.length !== raw.args.length) {
-    errors.push("args must contain only strings.");
-  }
-  if ("env" in raw && !isRecord(raw.env)) {
-    errors.push("env must be an object of environment variables.");
-  }
-  if (command && SHELL_OPERATOR_PATTERN.test(command)) {
-    errors.push("command must be an executable name/path, not a shell pipeline.");
-  }
-  for (const [index, arg] of args.entries()) {
-    if (SHELL_OPERATOR_PATTERN.test(arg)) {
-      warnings.push(`Argument contains shell-like syntax: ${sanitizedArgs[index] ?? "${SECRET}"}`);
-    }
-  }
-  const packageName = command ? packageFromRunner(command, args) : "";
-  const runnerName = command ? packageRunnerName(command, args) : "";
-  if (command) {
-    if (["npx", "uvx", "bunx"].includes(commandName(command)) && !packageName) {
-      errors.push(`${command} server is missing a package name in args.`);
-    }
-    if (packageName && runnerName && !isPinnedPackageSpec(packageName)) {
-      const parsed = parsePackageSpec(packageName);
-      const suggestedName = parsed?.name || packageName;
-      warnings.push(
-        `${runnerName} runs unpinned package ${packageName}; pin an exact version such as ${suggestedName}@1.2.3 before granting MCP tool access.`,
-      );
-    }
-  }
-  if (url) {
-    try {
-      const parsed = new URL(url);
-      const isLocal =
-        parsed.hostname === "localhost" ||
-        parsed.hostname === "127.0.0.1" ||
-        // URL.hostname returns IPv6 hosts wrapped in brackets, so the loopback
-        // address surfaces as "[::1]" (never the bare "::1").
-        parsed.hostname === "[::1]";
-      if (parsed.protocol !== "https:" && !(isLocal && parsed.protocol === "http:")) {
-        warnings.push("Remote MCP URLs should use HTTPS unless they are localhost.");
-      }
-    } catch {
-      errors.push("url must be a valid URL.");
-    }
-  }
-  for (const [key, value] of Object.entries(env)) {
-    const normalizedValue = String(value ?? "");
-    if (
-      SENSITIVE_ENV_PATTERN.test(key) &&
-      normalizedValue &&
-      !PLACEHOLDER_PATTERN.test(normalizedValue)
-    ) {
-      warnings.push(`${key} looks secret-like and was redacted from output.`);
-    }
-    if (PLACEHOLDER_PATTERN.test(normalizedValue)) {
-      warnings.push(`${key} is still a placeholder.`);
-    }
-  }
-
-  return {
-    name: normalizedName || name,
-    transport: url ? ("remote" as const) : command ? ("stdio" as const) : ("unknown" as const),
-    command,
-    url: typeof sanitizedRawValue.url === "string" ? sanitizedRawValue.url : url,
-    packageName: command ? packageFromRunner(command, sanitizedArgs) : "",
-    envKeys,
-    errors,
-    warnings: [...new Set(warnings)],
-    sanitized,
-    redactedSecretCount,
-  };
-}
-
-function buildReportText(result: Omit<McpConfigValidation, "reportText">) {
-  const lines = [
-    `MCP config validation: ${result.ok ? "pass" : "blocked"}`,
-    `Servers: ${result.servers.length}`,
-    `Errors: ${result.errors.length}`,
-    `Warnings: ${result.warnings.length}`,
-    `Redacted secrets: ${result.redactedSecretCount}`,
-  ];
-
-  for (const server of result.servers) {
-    lines.push(
-      "",
-      `## ${server.name}`,
-      `Transport: ${server.transport}`,
-      server.packageName ? `Package: ${server.packageName}` : "",
-      server.url ? `URL: ${server.url}` : "",
-      server.envKeys.length ? `Env keys: ${server.envKeys.join(", ")}` : "Env keys: none",
-    );
-    for (const error of server.errors) lines.push(`- Error: ${error}`);
-    for (const warning of server.warnings) lines.push(`- Warning: ${warning}`);
-  }
-
-  return lines.filter(Boolean).join("\n");
-}
+export type { McpConfigServerReport, McpConfigValidation } from "@/lib/mcp-config-validator-lib";
 
 export function validateMcpConfigText(input: string): McpConfigValidation {
   const errors: string[] = [];
