+++ tests/mcp-config-validator-lib.test.ts
@@ -1,0 +1,310 @@
+import { describe, expect, it } from "vitest";
+
+import {
+  buildReportText,
+  commandName,
+  extractServers,
+  firstPackageOperand,
+  isRecord,
+  packageFromRunner,
+  packageRunnerName,
+  redactArgValue,
+  redactEnvValue,
+  redactUrlValue,
+  sanitizeConfigValue,
+  splitArgPlaceholder,
+  validateServer,
+} from "../apps/web/src/lib/mcp-config-validator-lib";
+
+describe("MCP config validator lib", () => {
+  describe("redactEnvValue", () => {
+    it("redacts secret-like env keys and raw token values", () => {
+      expect(
+        redactEnvValue("GITHUB_TOKEN", "ghp_abcdefghijklmnopqrstuvwxyz123456"),
+      ).toBe("${GITHUB_TOKEN}");
+      expect(
+        redactEnvValue("owner", "ghp_abcdefghijklmnopqrstuvwxyz123456"),
+      ).toBe("${OWNER}");
+      expect(redactEnvValue("LOG_LEVEL", "debug")).toBe("debug");
+    });
+
+    it("preserves placeholders and empty values", () => {
+      expect(redactEnvValue("API_KEY", "${API_KEY}")).toBe("${API_KEY}");
+      expect(redactEnvValue("API_KEY", "")).toBe("");
+      expect(redactEnvValue("", "ghp_abcdefghijklmnopqrstuvwxyz123456")).toBe(
+        "${SECRET}",
+      );
+    });
+  });
+
+  describe("redactUrlValue", () => {
+    it("redacts credentials and sensitive query params", () => {
+      const result = redactUrlValue(
+        "https://user:password@example.com/mcp?api_key=sk-12345678901234567890&mode=read",
+      );
+      expect(result.redactedCount).toBe(1);
+      expect(result.value).toContain("${URL_USERNAME}");
+      expect(result.value).toContain("${URL_PASSWORD}");
+      expect(result.value).toContain("${API_KEY}");
+      expect(result.value).not.toContain("password");
+      expect(result.value).not.toContain("sk-12345678901234567890");
+    });
+
+    it("falls back to env redaction for malformed URLs", () => {
+      const result = redactUrlValue(
+        "https://bad host/sse?api_key=sk-12345678901234567890",
+      );
+      expect(result.redactedCount).toBe(1);
+      expect(result.value).toBe("${URL}");
+    });
+
+    it("preserves encoded placeholders that are not generated by redaction", () => {
+      const result = redactUrlValue(
+        "https://evil.example/mcp?x=%24%7BGITHUB_TOKEN%7D&safe=1",
+      );
+      expect(result.redactedCount).toBe(0);
+      expect(result.value).toContain("%24%7BGITHUB_TOKEN%7D");
+      expect(result.value).not.toContain("${GITHUB_TOKEN}");
+    });
+  });
+
+  describe("redactArgValue", () => {
+    it("redacts key=value args and bare token prefixes", () => {
+      expect(redactArgValue("token=ghp_12345678901234567890")).toEqual({
+        value: "token=${TOKEN}",
+        redactedCount: 1,
+      });
+      expect(redactArgValue("gho_abcdefghijklmnopqrstuvwxyz123456")).toEqual({
+        value: "${SECRET}",
+        redactedCount: 1,
+      });
+    });
+
+    it("delegates http(s) args to URL redaction", () => {
+      const result = redactArgValue(
+        "https://example.com/sse?authorization=Bearer%20abcdef1234567890",
+      );
+      expect(result.redactedCount).toBe(1);
+      expect(result.value).toContain("${AUTHORIZATION}");
+    });
+  });
+
+  describe("splitArgPlaceholder", () => {
+    it("maps sensitive CLI flags to placeholder keys", () => {
+      expect(splitArgPlaceholder("--api-key")).toBe("API_KEY");
+      expect(splitArgPlaceholder("--client-secret")).toBe("CLIENT_SECRET");
+      expect(splitArgPlaceholder("--mode")).toBe("");
+      expect(splitArgPlaceholder("--token=inline")).toBe("");
+    });
+  });
+
+  describe("sanitizeConfigValue", () => {
+    it("recursively redacts nested objects and header maps", () => {
+      const result = sanitizeConfigValue("root", {
+        headers: {
+          Authorization: "Bearer abcdefghijklmnopqrstuvwxyz123456",
+          "x-api-key": "sk-liveapikeyvaluedonotleak",
+        },
+        metadata: { owner: "docs" },
+      });
+      expect(result.redactedCount).toBe(2);
+      expect(result.value).toEqual({
+        headers: {
+          Authorization: "${AUTHORIZATION}",
+          "x-api-key": "${X_API_KEY}",
+        },
+        metadata: { owner: "docs" },
+      });
+    });
+
+    it("redacts split-flag arg pairs", () => {
+      const result = sanitizeConfigValue("args", [
+        "--api-key",
+        "live_api_key_value",
+        "--token",
+        "${EXISTING_TOKEN}",
+      ]);
+      expect(result.redactedCount).toBe(1);
+      expect(result.value).toEqual([
+        "--api-key",
+        "${API_KEY}",
+        "--token",
+        "${EXISTING_TOKEN}",
+      ]);
+    });
+  });
+
+  describe("package runner helpers", () => {
+    it("normalizes windows executables and extracts package operands", () => {
+      expect(commandName("C:\\tools\\npx.cmd")).toBe("npx");
+      expect(commandName("uvx.exe")).toBe("uvx");
+      expect(firstPackageOperand(["-y", "--", "@scope/pkg"])).toBe(
+        "@scope/pkg",
+      );
+      expect(
+        packageFromRunner("npm", ["exec", "my-package", "--", "--verbose"]),
+      ).toBe("my-package");
+      expect(
+        packageFromRunner("docker", ["run", "ghcr.io/example/mcp:1.2.3"]),
+      ).toBe("ghcr.io/example/mcp:1.2.3");
+    });
+
+    it("labels package runners for unpinned warnings", () => {
+      expect(packageRunnerName("pnpm", ["dlx", "@example/mcp"])).toBe(
+        "pnpm dlx",
+      );
+      expect(packageRunnerName("npm", ["exec", "@scope/pkg"])).toBe("npm exec");
+      expect(packageRunnerName("npx", ["-y", "@example/mcp"])).toBe("npx");
+      expect(packageRunnerName("docker", ["run", "image"])).toBe("");
+    });
+  });
+
+  describe("extractServers", () => {
+    it("accepts wrapped and bare server maps", () => {
+      expect(
+        extractServers({ mcpServers: { demo: { command: "npx" } } }),
+      ).toEqual({
+        servers: { demo: { command: "npx" } },
+        wrapped: false,
+      });
+      expect(
+        extractServers({ demo: { url: "https://example.com/mcp" } }),
+      ).toEqual({
+        servers: { demo: { url: "https://example.com/mcp" } },
+        wrapped: true,
+      });
+    });
+
+    it("rejects non-objects and empty configs", () => {
+      expect(extractServers(null)).toEqual({
+        servers: {},
+        rootError: "Config must be a JSON object.",
+      });
+      expect(extractServers({})).toEqual({
+        servers: {},
+        rootError: "Config must include an mcpServers object.",
+      });
+    });
+  });
+
+  describe("validateServer", () => {
+    it("classifies stdio and remote transports with env metadata", () => {
+      const stdio = validateServer("github", {
+        command: "npx",
+        args: ["-y", "@modelcontextprotocol/server-github"],
+        env: { GITHUB_PERSONAL_ACCESS_TOKEN: "ghp_real_token_value" },
+      });
+      expect(stdio).toMatchObject({
+        transport: "stdio",
+        packageName: "@modelcontextprotocol/server-github",
+        envKeys: ["GITHUB_PERSONAL_ACCESS_TOKEN"],
+        errors: [],
+      });
+      expect(stdio.redactedSecretCount).toBe(1);
+      expect(stdio.sanitized.env).toEqual({
+        GITHUB_PERSONAL_ACCESS_TOKEN: "${GITHUB_PERSONAL_ACCESS_TOKEN}",
+      });
+
+      const remote = validateServer("local", { url: "http://[::1]:3000/sse" });
+      expect(remote.transport).toBe("remote");
+      expect(remote.warnings).not.toContain(
+        "Remote MCP URLs should use HTTPS unless they are localhost.",
+      );
+    });
+
+    it("flags invalid names, missing transport, and shell pipelines", () => {
+      const invalid = validateServer("../bad", {
+        command: "npx && rm -rf /",
+        args: ["-y"],
+      });
+      expect(invalid.errors).toEqual(
+        expect.arrayContaining([
+          "Server name must be 1-64 characters of letters, numbers, dot, underscore, or dash.",
+          "command must be an executable name/path, not a shell pipeline.",
+        ]),
+      );
+
+      const missing = validateServer("empty", {});
+      expect(missing.errors).toContain(
+        "Server must define command for stdio or url for remote transport.",
+      );
+      expect(missing.transport).toBe("unknown");
+    });
+
+    it("warns on unpinned packages and remote http URLs", () => {
+      const report = validateServer("bare", {
+        command: "npx",
+        args: ["-y", "@modelcontextprotocol/server-github"],
+      });
+      expect(report.warnings.join("\n")).toContain("runs unpinned package");
+
+      const remote = validateServer("remote", {
+        url: "http://example.com:3000/sse",
+      });
+      expect(remote.warnings).toContain(
+        "Remote MCP URLs should use HTTPS unless they are localhost.",
+      );
+    });
+
+    it("rejects non-object server configs", () => {
+      expect(validateServer("broken", "stdio")).toMatchObject({
+        transport: "unknown",
+        errors: ["Server config must be an object."],
+      });
+    });
+  });
+
+  describe("buildReportText", () => {
+    it("summarizes validation output for pass and blocked runs", () => {
+      const pass = buildReportText({
+        ok: true,
+        errors: [],
+        warnings: [],
+        servers: [
+          {
+            name: "github",
+            transport: "stdio",
+            packageName: "@example/mcp",
+            envKeys: ["TOKEN"],
+            errors: [],
+            warnings: [],
+          },
+        ],
+        fixedConfigText: "{}",
+        redactedSecretCount: 2,
+      });
+      expect(pass).toContain("MCP config validation: pass");
+      expect(pass).toContain("Redacted secrets: 2");
+      expect(pass).toContain("## github");
+      expect(pass).toContain("Package: @example/mcp");
+
+      const blocked = buildReportText({
+        ok: false,
+        errors: ["demo: url must be a valid URL."],
+        warnings: ["demo: TOKEN is still a placeholder."],
+        servers: [
+          {
+            name: "demo",
+            transport: "remote",
+            envKeys: [],
+            errors: ["url must be a valid URL."],
+            warnings: ["TOKEN is still a placeholder."],
+          },
+        ],
+        fixedConfigText: "",
+        redactedSecretCount: 0,
+      });
+      expect(blocked).toContain("MCP config validation: blocked");
+      expect(blocked).toContain("- Error: url must be a valid URL.");
+      expect(blocked).toContain("- Warning: TOKEN is still a placeholder.");
+    });
+  });
+
+  describe("isRecord", () => {
+    it("narrows plain objects", () => {
+      expect(isRecord({})).toBe(true);
+      expect(isRecord([])).toBe(false);
+      expect(isRecord(null)).toBe(false);
+    });
+  });
+});
